+++ src/main/java/com/dace/dmgr/combat/character/jager/action/JagerUlt.java
@@ -278,7 +278,7 @@ public final class JagerUlt extends UltimateSkill {
         }
 
         @Override
-        public void onAttack(@NonNull Damageable victim, int damage, @NonNull DamageType damageType, boolean isCrit, boolean isUlt) {
+        public void onAttack(@NonNull Damageable victim, double damage, @NonNull DamageType damageType, boolean isCrit, boolean isUlt) {
             owner.onAttack(victim, damage, damageType, isCrit, isUlt);
 
             if (victim instanceof CombatUser)
@@ -291,7 +291,7 @@ public final class JagerUlt extends UltimateSkill {
         }
 
         @Override
-        public void onDamage(@Nullable Attacker attacker, int damage, int reducedDamage, @NonNull DamageType damageType, @Nullable Location location,
+        public void onDamage(@Nullable Attacker attacker, double damage, double reducedDamage, @NonNull DamageType damageType, @Nullable Location location,
                              boolean isCrit, boolean isUlt) {
             SoundUtil.playNamedSound(NamedSound.COMBAT_JAGER_ULT_DAMAGE, entity.getLocation(), 1 + damage * 0.001);
             CombatEffectUtil.playBreakEffect(location, entity, damage);
