+++ src/main/java/com/dace/dmgr/combat/character/silia/Silia.java
@@ -129,7 +129,7 @@ public final class Silia extends Scuffler {
     }
 
     @Override
-    public boolean onAttack(@NonNull CombatUser attacker, @NonNull Damageable victim, int damage, @NonNull DamageType damageType, boolean isCrit) {
+    public boolean onAttack(@NonNull CombatUser attacker, @NonNull Damageable victim, double damage, @NonNull DamageType damageType, boolean isCrit) {
         if (victim instanceof CombatUser && isCrit)
             attacker.addScore("백어택", SiliaT1Info.CRIT_SCORE);
 
@@ -137,7 +137,7 @@ public final class Silia extends Scuffler {
     }
 
     @Override
-    public void onDamage(@NonNull CombatUser victim, @Nullable Attacker attacker, int damage, @NonNull DamageType damageType, Location location, boolean isCrit) {
+    public void onDamage(@NonNull CombatUser victim, @Nullable Attacker attacker, double damage, @NonNull DamageType damageType, Location location, boolean isCrit) {
         CombatEffectUtil.playBleedingEffect(location, victim.getEntity(), damage);
     }
