+++ src/main/java/com/dace/dmgr/combat/entity/temporary/Barrier.java
@@ -56,7 +56,7 @@ public abstract class Barrier<T extends Entity> extends SummonEntity<T> implemen
 
     @Override
     @MustBeInvokedByOverriders
-    public void onDamage(@Nullable Attacker attacker, int damage, int reducedDamage, @NonNull DamageType damageType, @Nullable Location location,
+    public void onDamage(@Nullable Attacker attacker, double damage, double reducedDamage, @NonNull DamageType damageType, @Nullable Location location,
                          boolean isCrit, boolean isUlt) {
         if (owner.getGameUser() != null)
             owner.getGameUser().setDefend(owner.getGameUser().getDefend() + damage);
